requirements.txt
# Core encoding/decoding
numpy
Pillow
imageio-ffmpeg
tqdm
yt-dlp
requests
cryptography
reedsolo

# GUI (desktop app)
ttkbootstrap

# Web API
flask>=3.0
flask-cors

# Google OAuth + YouTube API (optional — needed for /api/upload/oauth/start)
google-auth
google-auth-oauthlib
google-api-python-client